removed typecast

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -12,6 +12,6 @@ int set_bit(unsigned long int *n, unsigned int index)
 	if (index >= sizeof(unsigned long int) * 8)
 		return (-1);
 
-	*n |= ((long)1 << index);
+	*n |= (1 << index);
 	return (0);
 }
